Moved simulated context to simulator
